Research Artefak: Implementation of Hyperparameter Tuning on Bert, Albert, And Gpt Algorithms For Classification of React-Related Errors in Stack Overflow
This is a research artifact for the paper: IMPLEMENTASI HYPERPARAMETER TUNING PADA ALGORITMA BERT, ALBERT, DAN GPT UNTUK KLASIFIKASI EROR TERKAIT REACT DI STACK OVERFLOW. This artifact is a repository consisting of a collected dataset including (i) a total of 583,419 Stack Overflow questions specifically related to React errors, (ii) 1,533 manually labeled data samples. The purposes of this artifact are to enable researchers to replicate our quantitative experimental results presented in the paper, and to reuse the dataset and tuning configurations for further natural language processing and software engineering research.